updated steps
